git-svn-id: file:///svn/toku/tokudb.1032b@7835 c7de825b-a66e-492c-adef-691d508d4ae1

+++ src/tests/test_db_curs2.c
@@ -1,12 +1,10 @@
 /* Primary with two associated things. */
 
-#include <arpa/inet.h>
 #include <assert.h>
 #include <db.h>
 #include <errno.h>
 #include <string.h>
 #include <sys/stat.h>
-#include <sys/time.h>
 #include <unistd.h>
 #include <ctype.h>
 
@@ -122,13 +120,13 @@ read_uchar_from_dbt (const DBT *dbt, unsigned int *off, unsigned char *uchar) {
 }
 
 static void
-read_uint_from_dbt (const DBT *dbt, unsigned int *off, unsigned int *uint) {
+read_uint_from_dbt (const DBT *dbt, unsigned int *off, unsigned int *uintptr) {
     unsigned char a,b,c,d;
     read_uchar_from_dbt(dbt, off, &a);
     read_uchar_from_dbt(dbt, off, &b);
     read_uchar_from_dbt(dbt, off, &c);
     read_uchar_from_dbt(dbt, off, &d);
-    *uint = (a<<24)+(b<<16)+(c<<8)+d;
+    *uintptr = (a<<24)+(b<<16)+(c<<8)+d;
 }
 
 static void
