+++ src/tests/test_txn_abort5a.c
@@ -5,7 +5,6 @@
 #include <string.h>
 #include <sys/stat.h>
 #include <db.h>
-#include <arpa/inet.h>
 #include "test.h"
 
 
@@ -66,7 +65,7 @@ test_txn_abort (int n) {
     for (i=0; 1; i++) {
 	r = cursor->c_get(cursor, &key, &val, DB_NEXT);
 	if (r!=0) break;
-	if (verbose>2) printf("%u present\n", ntohl(*(int*)key.data));
+	if (verbose>2) printf("%u present\n", (uint32_t)ntohl(*(int*)key.data));
 	assert(key.size==4);
 	assert(ntohl(*(int*)key.data)==(unsigned int)(2*i));
     }
